Accept SAML assertions from Shibboleth IdPs.
Shibboleth IdPs encode their assertions using different xml tags. We either need to make the xpath searches namespace aware (which runs the risk of not 'working' when other SAML providers don't correctly declare the namespaces) or modify the code to work with assertions made by the standard Shibboleth libraries.

This change looks first for saml and samlp tags, and if not found tries saml2 and saml2p.