fix verifying signatures and decryption
Before we were erroring out if we couldn't find a Signature in a Response,
however, the signature might be inside the encrypted portion. Now we will
check for a signature before decryption, then again after decrypting (if we
didn't find on the first time). We only fail if no signatures are found at
all.

This is another band-aid around the fact that we're not using a real xmlsec
library.