guard against duplicate id XML signature envelope attacks
because we have to use xmlSecAddIDs to get xmlsec to be able to
dereference signature references in SAML responses, a specially crafted
document with duplicate IDs can fool xmlsec into validating a signature
outside of the ancestry of the reference. this provides an attack vector
for injecting an unsigned assertion.

the solution is to catch the error thrown when xmlsec sees the duplicate
ID and immediately treat the response as invalid when this occurs.